implemented bag places; always visible on map. some code is unnecessary ATM, but is important for later functions
