+++ frontend/src/components/Map.scss
@@ -1,0 +1,21 @@
+.locationButton{
+  position: relative;
+  top: 0px;
+  width: 164px;
+  left: 50%;
+  -ms-transform: translate(-50%, -50%);
+  transform: translate(-50%, -50%);
+
+}
+
+.markerButton{
+
+  //top: 20px;
+  //width: 150px;
+
+  top: 38px;
+  left: 50%;
+  -ms-transform: translate(-50%, -50%);
+  transform: translate(-50%, -50%);
+
+}
