--- frontend/src/models/LocationData.ts
@@ -1,4 +1,0 @@
-export interface LocationData {
-    latitude: string;
-    longitude: string;
-}
